Amended the enum to be a concern as it needs to be shared in the model
