+++ app/models/concerns/region_code.rb
@@ -1,0 +1,23 @@
+module RegionCode
+  extend ActiveSupport::Concern
+  included do
+    # These correspond to the first-level NUTS regions for the UK (minus Northern Ireland)
+    # https://en.wikipedia.org/wiki/First-level_NUTS_of_the_European_Union#United_Kingdom
+
+    enum region_code: {
+      'No region' => 0,
+      'London' => 1,
+      'South East' => 2,
+      'South West' => 3,
+      'Wales' => 4,
+      'West Midlands' => 5,
+      'East Midlands' => 6,
+      'Eastern' => 7,
+      'North West' => 8,
+      'Yorkshire & the Humber' => 9,
+      'North East' => 10,
+      'Scotland' => 11,
+    }
+
+  end
+end
