+++ spec/serializers/site_serializer_spec.rb
@@ -17,7 +17,28 @@ require "rails_helper"
 
 RSpec.describe SiteSerializer do
   let(:site) { create :site }
+
   subject { serialize(site) }
 
-  it { is_expected.to include(name: site.location_name, campus_code: site.code, region_code: site.region_code) }
+  it { is_expected.to include(name: site.location_name, campus_code: site.code) }
+  it { is_expected.to include(region_code: site.region_code) }
+end
+
+
+RSpec.describe SiteSerializer do
+  subject do
+    serialize(site)["region_code"]
+  end
+
+  region_codes = 1..11
+
+  region_codes.each do |region_code|
+    describe "region code #{region_code} " do
+
+      let(:site) { create :site, region_code: region_code}
+      it {is_expected.to eql(format("%02d", region_code))}
+      it {expect(subject.length).to eql(2)}
+
+    end
+  end
 end
